refactor(frontend): rename ResetPassword
